Pathfinding from the current location
Instead of recalculating the path form the start location (which caused enemies to jump to the next tile) calculate from the current location.

Pass the location as a parameter into the script and calculate the new path.

Stop the current Coroutine with StopAllCoroutines, calculate the new path then start a new Coroutine to follow the path.